Refactor logout button to be own component
