+++ client/src/components/Fridge.js
@@ -1,43 +1,16 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
 import { StackNavigator } from 'react-navigation';
-import {
-  Button,
-  View,
-  Image,
-  Text,
-} from 'react-native'
+import { Image } from 'react-native'
 
-import * as authActions from '../actions/authActions.js';
 import FridgeView from './FridgeView'
 import CategoryView from './CategoryView'
 import AddItem from './AddItem'
-
-let Logout = (props) => (
-  <Button title='Logout' onPress={() => {
-    props.authActions.logoutUser()
-  }}/>
-);
-
-const LogoutState = (store) => {
-  return {
-    fridgeName: store.fridge.fridge.name
-  }
-};
-
-const LogoutDispatch = (dispatch) => {
-  return {
-    authActions: bindActionCreators(authActions, dispatch)
-  }
-};
-
-Logout = connect(LogoutState, LogoutDispatch)(Logout);
+import Logout from './Logout';
 
 const Logo = (
   <Image 
     source={require('../images/logo.png')} 
-    style={{height: 40, width: 40}} 
+    style={{height: 40, width: 40, paddingBottom: 2}} 
   />
 );
 
@@ -49,9 +22,7 @@ const Fridge = StackNavigator({
   initialRouteParams: {
     something: 'something'
   },
-  navigationOptions: (props) => {
-    console.log(props);
-    return ({
+  navigationOptions: (props) => ({
       headerTitle: Logo,
       headerRight: (<Logout />),
       headerStyle: {
@@ -60,6 +31,6 @@ const Fridge = StackNavigator({
       }
     })
   },
-});
+);
 
 export default Fridge;
